begin updated layout
add a couple images for logo, the main grid background, and build out the header.  renames index.html in /pubic to index.html.off, the view i'm adding the layout to is the previous wireframey view in /exteriors/index. comments out all yields in app layout.

viewing root path at this point should just show a page with the grid background and the header.